+++ config/messages/app_ru.properties
@@ -3,7 +3,7 @@ app.footer=Приложение по вебинару <a href\="https\://github.
 app.login=Зайти как
 
 users.title=Список пользователей
-users.edit=Радактировать пользователя
+users.edit=Редактировать пользователя
 
 meals.title=Список еды
 meals.edit=Редактировать еду
