--- src/main/java/ru/javawebinar/topjava/web/MealServlet.java
@@ -1,97 +1,0 @@
-package ru.javawebinar.topjava.web;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.context.support.WebApplicationContextUtils;
-import ru.javawebinar.topjava.model.UserMeal;
-import ru.javawebinar.topjava.util.TimeUtil;
-import ru.javawebinar.topjava.web.meal.UserMealRestController;
-
-import javax.servlet.ServletConfig;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.util.Objects;
-
-/**
- * User: gkislin
- * Date: 19.08.2014
- */
-public class MealServlet extends HttpServlet {
-    private static final Logger LOG = LoggerFactory.getLogger(MealServlet.class);
-
-    private UserMealRestController mealController;
-
-    @Override
-    public void init(ServletConfig config) throws ServletException {
-        super.init(config);
-        WebApplicationContext springContext = WebApplicationContextUtils.getRequiredWebApplicationContext(getServletContext());
-        mealController = springContext.getBean(UserMealRestController.class);
-    }
-
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.setCharacterEncoding("UTF-8");
-        String action = request.getParameter("action");
-        if (action == null) {
-            final UserMeal userMeal = new UserMeal(
-                    LocalDateTime.parse(request.getParameter("dateTime")),
-                    request.getParameter("description"),
-                    Integer.valueOf(request.getParameter("calories")));
-
-            if (request.getParameter("id").isEmpty()) {
-                LOG.info("Create {}", userMeal);
-                mealController.create(userMeal);
-            } else {
-                LOG.info("Update {}", userMeal);
-                mealController.update(userMeal, getId(request));
-            }
-            response.sendRedirect("meals");
-
-        } else if (action.equals("filter")) {
-            LocalDate startDate = TimeUtil.parseLocalDate(resetParam("startDate", request));
-            LocalDate endDate = TimeUtil.parseLocalDate(resetParam("endDate", request));
-            LocalTime startTime = TimeUtil.parseLocalTime(resetParam("startTime", request));
-            LocalTime endTime = TimeUtil.parseLocalTime(resetParam("endTime", request));
-            request.setAttribute("mealList", mealController.getBetween(startDate, startTime, endDate, endTime));
-            request.getRequestDispatcher("/mealList.jsp").forward(request, response);
-        }
-    }
-
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String action = request.getParameter("action");
-
-        if (action == null) {
-            LOG.info("getAll");
-            request.setAttribute("mealList", mealController.getAll());
-            request.getRequestDispatcher("mealList.jsp").forward(request, response);
-        } else if ("delete".equals(action)) {
-            int id = getId(request);
-            LOG.info("Delete {}", id);
-            mealController.delete(id);
-            response.sendRedirect("meals");
-        } else {
-            final UserMeal meal = "create".equals(action) ?
-                    new UserMeal(LocalDateTime.now(), "", 1000) :   // create
-                    mealController.get(getId(request));             // update
-            request.setAttribute("meal", meal);
-            request.getRequestDispatcher("mealEdit.jsp").forward(request, response);
-        }
-    }
-
-    private String resetParam(String param, HttpServletRequest request) {
-        String value = request.getParameter(param);
-        request.setAttribute(param, value);
-        return value;
-    }
-
-    private int getId(HttpServletRequest request) {
-        String paramId = Objects.requireNonNull(request.getParameter("id"), "parameter id  must not be null");
-        return Integer.valueOf(paramId);
-    }
-}
