--- src/main/java/ru/javawebinar/topjava/web/UserServlet.java
@@ -1,46 +1,0 @@
-package ru.javawebinar.topjava.web;
-
-import org.slf4j.Logger;
-import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.context.support.WebApplicationContextUtils;
-import ru.javawebinar.topjava.LoggedUser;
-import ru.javawebinar.topjava.web.user.AdminRestController;
-
-import javax.servlet.ServletConfig;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-import static org.slf4j.LoggerFactory.getLogger;
-
-/**
- * User: gkislin
- * Date: 19.08.2014
- */
-public class UserServlet extends HttpServlet {
-    private static final Logger LOG = getLogger(UserServlet.class);
-
-    private AdminRestController adminController;
-
-    @Override
-    public void init(ServletConfig config) throws ServletException {
-        super.init(config);
-        WebApplicationContext springContext = WebApplicationContextUtils.getRequiredWebApplicationContext(getServletContext());
-        adminController = springContext.getBean(AdminRestController.class);
-    }
-
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        int userId = Integer.valueOf(request.getParameter("userId"));
-        LoggedUser.setId(userId);
-        response.sendRedirect("meals");
-    }
-
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        LOG.debug("getAll");
-        request.setAttribute("userList", adminController.getAll());
-        request.getRequestDispatcher("/userList.jsp").forward(request, response);
-//        response.sendRedirect("userList.jsp");
-    }
-}
